requirements.txt
streamlit>=1.28.0
torch>=1.13.0
torchvision>=0.14.0
ultralytics>=8.0.0
easyocr>=1.6.0
opencv-python-headless>=4.6.0
Pillow>=9.2.0
matplotlib>=3.5.0
numpy>=1.21.0